requirements.txt
discord.py
python-dotenv
requests
animec
peewee
